Support Rails 5.2
Add example app and add it to the build to see if any additional changes
are required.

Need to require the MessageEncryptor class as Rails errors if it's not
required. This may be a bug? The MessageVerifier is loaded by Rails in
the same file as would raise an "uninitialized constant" error.

```
Failure/Error: require "action_dispatch/middleware/session/abstract_store"

NameError:
  uninitialized constant ActiveSupport::MessageEncryptor
  Did you mean?  ActiveSupport::MessageVerifier
```

Requiring it manually in spec fixes the test suite for Rails 5.2.
Running the test app manually works fine, something is off in the load
order.