Move CI to GitHub Actions
In the process I've removed old Rails 4.0 versions because they're long
EOL and I can't make it work locally or on the CI.

It doesn't test on all the EOL Ruby versions, but only tests on the
Ruby 2.6 or 2.7 depending on the Rails version it tests agains.

I've not migrated danger as there's another PR open for that.